Quaternionic functions

From AbstractQuaternion{T} we define three subtypes:

Quaternion{T}, which is an element of the general algebra of quaternions over any T<:Real.

Rotor{T}, which is an element of the multiplicative group of unit quaternions, and is interpreted as mapping to a rotation. The magnitude is assumed to be 1 (though, for efficiency, this is not generally confirmed), and the sign may be freely changed in certain cases.

QuatVec{T}, which is an element of the additive group of quaternions with 0 scalar part; a "pure vector" quaternion.

For simplicity, almost every function in this package is defined for general Quaternions, so you may not need any other type. However, it can frequently be more accurate and more efficient to use the other subtypes where relevant.

Constructors, constants, and conversions

At the most basic level, Quaternion{T} mimics Complex{T} as closely as possible, including the behavior of most functions in Base. The Rotor{T} and QuatVec{T} subtypes behave very similarly, except that most of their constructors automatically impose the constraints that the norm is 1 and the scalar component is 0, respectively. Also note that when a certain operation is not defined for either of those subtypes, the functions will usually convert to a general Quaternion automatically.

To create new Quaternions interactively, it is typically most convenient to use the constants imx, imy, and imz — or equivalently 𝐢, 𝐣, and 𝐤 — multiplied by appropriate factors and added together. For programmatic work, it is more common to use the Quaternion function — which takes all four components, the three vector components, or just the one scalar component, and creates a new Quaternion of the type implied by the arguments. You can also specify the type, as in Quaternion{Float64}(...). Type conversions with promote, widen, float, etc., work as expected.

Quaternionic.QuatVec — Type

QuatVec{T<:Number} <: Number

Pure-vector quaternion with elements of type T. These objects can be significantly faster and more accurate in certain operations than general Quaternions.

QuatVecF16, QuatVecF32 and QuatVecF64 are aliases for QuatVec{Float16}, QuatVec{Float32} and QuatVec{Float64} respectively. See also Quaternion and Rotor.

The functions

quatvec(w, x, y, z)
quatvec(x, y, z)
quatvec(w)

create a new QuatVec with the given components (where the components are as described in Quaternion), except that the scalar argument w is always set to 0.

Examples

julia> quatvec(1, 2, 3, 4)
 + 2𝐢 + 3𝐣 + 4𝐤
julia> quatvec(quaternion(1, 2, 3, 4))
 + 2𝐢 + 3𝐣 + 4𝐤
julia> quatvec(2, 3, 4)
 + 2𝐢 + 3𝐣 + 4𝐤
julia> quatvec(1)
 + 0𝐢 + 0𝐣 + 0𝐤

Quaternionic.Quaternion — Type

Quaternion{T<:Number} <: Number

Quaternionic number type with elements of type T.

QuaternionF16, QuaternionF32 and QuaternionF64 are aliases for Quaternion{Float16}, Quaternion{Float32} and Quaternion{Float64} respectively. See also Rotor and QuatVec.

The functions

quaternion(w, x, y, z)
quaternion(x, y, z)
quaternion(w)

create a new quaternion with the given components. The argument w is the scalar component, and x, y, and z are the corresponding "vector" components. If any of these arguments is missing, it will be set to zero. The type of the returned quaternion will be inferred from the input arguments, or can be specified, by passing the type parameter T as above.

Note that the constants imx, imy, and imz can also be used like the complex im to create new Quaternion object.

Examples

julia> quaternion(1, 2, 3, 4)
1 + 2𝐢 + 3𝐣 + 4𝐤
julia> Quaternion{Float64}(1, 2, 3, 4)
1.0 + 2.0𝐢 + 3.0𝐣 + 4.0𝐤
julia> quaternion(1.0, 2.0, 3.0, 4.0)
1.0 + 2.0𝐢 + 3.0𝐣 + 4.0𝐤
julia> quaternion(2, 3, 4)
0 + 2𝐢 + 3𝐣 + 4𝐤
julia> quaternion(1)
1 + 0𝐢 + 0𝐣 + 0𝐤

Quaternionic.Rotor — Type

Rotor{T<:Number} <: Number

Quaternion of unit magnitude with elements of type T. These objects can be significantly faster and more accurate in certain operations representing rotations.

A rotor is typically considered to be an element of the group $\mathrm{Spin}(3) ≃ \mathrm{SU}(2)$, which can be thought of as the subgroup of quaternions with norm 1. They are particularly useful as representations of rotations because a rotor $R$ acts on a vector $\vec{v}$ by "conjugation" as

\[\vec{v}' = R\, \vec{v}\, R^{-1}.\]

(which can be represented in code as R * v / R or, more efficiently, as R(v)). This operation preserves the inner product between any two vectors conjugated in this way, and so is a rotation. Note that, because there are two factors of $R$ here, the sign of $R$ does not affect the result. Therefore, $\mathrm{Spin}(3)$ forms a double cover of the rotation group $\mathrm{SO}(3)$. For this reason, it will occasionally be useful to disregard or arbitrarily change the sign of a Rotor (as in distance functions) — though this is not generally the default, and may cause problems if the input rotors change sign when the corresponding rotations are not so different (cf. unflip).

RotorF16, RotorF32 and RotorF64 are aliases for Rotor{Float16}, Rotor{Float32} and Rotor{Float64} respectively. See also Quaternion and QuatVec.

The functions

rotor(w, x, y, z)
rotor(w)

create a new rotor with the given components (where the components are as described in Quaternion), automatically normalizing them on input. Note that this normalization step is the key difference between the Rotor and rotor functions; if you would like to bypass normalization, you can call

Rotor{T}(w, x, y, z)
Rotor{T}(w)

in the same way as rotor, and w, x, y, z will be converted to type T. Alternatively, you can call

Rotor{T}(v)

where v<:AbstractArray can be converted to an SVector{4, T}. If you want to handle the normalization step, you can use LinearAlgebra.normalize.

However, once a Rotor is created, its norm will often be assumed to be precisely 1. So if its true norm is significantly different, you will likely see weird results — including vectors with very different lengths after "rotation" by a non-unit Rotor.

Note that simply creating a Quaternion that happens to have norm 1 does not make it a Rotor. However, you can pass such a Quaternion to the rotor function and get the desired result.

Examples

julia> rotor(1, 2, 3, 4)
rotor(0.18257418583505536 + 0.3651483716701107𝐢 + 0.5477225575051661𝐣 + 0.7302967433402214𝐤)
julia> rotor(quaternion(1, 2, 3, 4))
rotor(0.18257418583505536 + 0.3651483716701107𝐢 + 0.5477225575051661𝐣 + 0.7302967433402214𝐤)
julia> Rotor{Float16}(1, 2, 3, 4)
rotor(1.0 + 2.0𝐢 + 3.0𝐣 + 4.0𝐤)
julia> rotor(Rotor{Float16}(1, 2, 3, 4))
rotor(0.1826 + 0.3652𝐢 + 0.548𝐣 + 0.7305𝐤)
julia> rotor(1.0)
rotor(1.0 + 0.0𝐢 + 0.0𝐣 + 0.0𝐤)

Quaternionic.imx — Constant

imx
𝐢

The quaternionic unit associated with rotation about the x axis. Can also be entered as Unicode bold 𝐢 (which can be input as \bfi<tab>).

Note that — just as im is a Complex{Bool} — imx is a QuatVec{Bool}, and as soon as you multiply by a scalar of any other number type (e.g., a Float64) it will be promoted to a QuatVec of that number type, and once you add a scalar it will be promoted to a Quaternion.

See also imy and imz.

Examples

julia> imx * imx
-1 + 0𝐢 + 0𝐣 + 0𝐤
julia> 1.2imx
 + 1.2𝐢 + 0.0𝐣 + 0.0𝐤
julia> 1.2 + 3.4imx
1.2 + 3.4𝐢 + 0.0𝐣 + 0.0𝐤
julia> 1.2 + 3.4𝐢
1.2 + 3.4𝐢 + 0.0𝐣 + 0.0𝐤

Quaternionic.imy — Constant

imy
𝐣

The quaternionic unit associated with rotation about the y axis. Can also be entered as Unicode bold 𝐣 (which can be input as \bfj<tab>).

Note that — just as im is a Complex{Bool} — imy is a QuatVec{Bool}, and as soon as you multiply by a scalar of any other number type (e.g., a Float64) it will be promoted to a QuatVec of that number type, and once you add a scalar it will be promoted to a Quaternion.

See also imx and imz.

Examples

julia> imy * imy
-1 + 0𝐢 + 0𝐣 + 0𝐤
julia> 1.2imy
 + 0.0𝐢 + 1.2𝐣 + 0.0𝐤
julia> 1.2 + 3.4imy
1.2 + 0.0𝐢 + 3.4𝐣 + 0.0𝐤
julia> 1.2 + 3.4𝐣
1.2 + 0.0𝐢 + 3.4𝐣 + 0.0𝐤

Quaternionic.imz — Constant

imz
𝐤

The quaternionic unit associated with rotation about the z axis. Can also be entered as Unicode bold 𝐤 (which can be input as \bfk<tab>).

Note that — just as im is a Complex{Bool} — imz is a QuatVec{Bool}, and as soon as you multiply by a scalar of any other number type (e.g., a Float64) it will be promoted to a QuatVec of that number type, and once you add a scalar it will be promoted to a Quaternion.

See also imx and imy.

Examples

julia> imz * imz
-1 + 0𝐢 + 0𝐣 + 0𝐤
julia> 1.2imz
 + 0.0𝐢 + 0.0𝐣 + 1.2𝐤
julia> 1.2 + 3.4imz
1.2 + 0.0𝐢 + 0.0𝐣 + 3.4𝐤
julia> 1.2 + 3.4𝐤
1.2 + 0.0𝐢 + 0.0𝐣 + 3.4𝐤

Quaternionic.components — Method

components(q::AbstractQuaternion{T})

Return the components of q as stored in the struct — as an SVector{4, T}. The components are ordered as (w, x, y, z), where w is the scalar part and x, y, and z are the vector parts.

Number functions from Base

The standard Number functions that work for Complex, such as isfinite, iszero, etc., should work analogously for Quaternion. The hash, read, and write functions are also implemented. As noted in the Examples, broadcasting to each component is also implemented via broadcasted.

Quaternionic.iszerovalue — Method

iszerovalue(x)

This is essentially Base.iszero, but intended to be overridden for types like ForwardDiff.Dual, where the value may be zero, but if the tangent is not zero, then Base.iszero will return false. This function, on the other hand, ignores the tangent part, and will return true if and only if the value part is zero.

This is needed internally in math functions like exp, log, and sqrt, where we frequently need to switch the algorithm based on whether some components are zero. In those isolated cases, a Taylor series should be provided that will be exactly zero for, e.g., Float64, but will also work correctly for ForwardDiff.Dual and other ADs.

Like Base.iszero, this function is defined recursively for arrays and quaternions.

Quaternionic.value — Method

value(x)

This is essentially the identity function, but intended to be overridden for types like ForwardDiff.Dual, where the value is the part of the dual number that corresponds to the original function value, and the tangent part is the part that corresponds to the derivative.

Algebraic and other mathematical functions

Along with the basic binary operators, the essential mathematical functions like conj, abs, abs2, exp, log, etc., are implemented. Most of these functions are found in the Base module, and are simply overloaded methods of functions that should also be familiar from Complex types. Note that we use a slightly different interpretation of angle for Quaternion, compared to Complex. We also have absvec and abs2vec, which are not useful in a Complex context, but compute the relevant quantities for the "vector" component of a Quaternion.

Base.conj — Method

conj(q)

Return the quaternion conjugate, which flips the sign of each "vector" component.

Examples

julia> conj(quaternion(1,2,3,4))
1 - 2𝐢 - 3𝐣 - 4𝐤

LinearAlgebra.:⋅ — Method

p ⋅ q

Evaluate the inner ("dot") product between two quaternions. Equal to the scalar part of p * conj(q).

Note that this function is not very commonly used, except as a quick way to determine whether the two quaternions are more anti-parallel than parallel, for functions like unflip.

This can be typed as \cdot<tab> in a Julia-aware editor.

LinearAlgebra.normalize — Method

normalize(q)

Return a copy of this quaternion, normalized.

Note that this returns the same type as the input quaternion. If you want to convert to a Rotor, just call rotor(q), which includes a normalization step.

This extends LinearAlgebra.normalize for quaternion types.

Quaternionic.:× — Method

a × b

Return the cross product of two pure-vector quaternions. Equal to ½ of the commutator product a*b-b*a.

This can be typed as \times<tab> in a Julia-aware editor.

Quaternionic.:×̂ — Method

a ×̂ b

Return the direction of the cross product between a and b; the normalized vector along a×b — unless the magnitude is zero, in which case the zero vector is returned.

This can be typed as \times<tab>\hat<tab> in a Julia-aware editor.

Base.:^ — Method

q ^ s
^(q, s)

Exponentiation operator, equivalent to $\exp(s \log(q))$.

When s is a real number, this is useful for natural "linear" interpolation/extrapolation through quaternion space, starting from q^0 = 1 and going directly to q^1 = q. Specifically, when q is a Rotor, this provides a geodesic on the unit 3-sphere going between those two points. For more general interpolations, see the slerp function.

Base.abs — Method

abs(q)

Square root of the sum of the squares of all four components of the quaternion.

This function uses Julia's built-in hypot function to avoid overflow and underflow.

Note that the result for a Rotor is identically 1, even if that is not true numerically.

Examples

julia> abs(quaternion(1,2,4,10))
11.0

Base.abs2 — Method

abs2(q)

Sum of the squares of all four components of the quaternion.

Note that the result for a Rotor is identically 1, even if that is not true numerically.

Examples

julia> abs2(quaternion(1,2,4,10))
121

Base.angle — Method

angle(q)

Phase angle in radians of the (spinorial) rotation represented by this quaternion.

Note that this may be different from your interpretation of the angle of a complex number in an important way. Because quaternions act on vectors by conjugation — as in q*v*conj(q) — there are two copies of q involved in that expression; in some sense, a quaternion acts "twice". Therefore, this angle may be twice what you expect from an analogy with complex numbers — depending on how you interpret the correspondence between complex numbers and quaternions.

Also, because quaternions are spinors, rotation by 2π is not the identity operation; only a full rotation through 4π is. Rotation by 2π changes the sign of the quaternion, which has no effect on the rotation of vectors, but does have an effect on rotations of more general objects (spinors). This issue of sign is important when considering the continuity of quaternionic functions (as in interpolations, differentiation, and integration, for example). Therefore, the angle returned by this function will be in the range [0, 2π], rather than the range [0, π] used for complex numbers. If you only care about the effects on vectors, and want to map the result of this function to the range [0, π], you can use θ -> min(θ, 2π-θ).

Examples

julia> θ=1.2;

julia> R=exp(θ * imz / 2);  # R*v*conj(R) rotates v by θ about the z axis

julia> angle(R)
1.2

julia> angle(exp(6.2 * imz / 2))  # Note that this is greater than π
6.2

Base.exp — Method

exp(q)

Exponential of a quaternion.

The exponential of a quaternion is defined as usual by its power series, which converges for all finite quaternions:

\[\exp(q) = \sum_{k=0}^\infty \frac{q^k}{k!}.\]

Derivatives at 0

Note that automatic differentiation of this function at exactly q=0 will only be accurate through fifth order. Using sixth-order derivatives or higher is so unusual that this will likely not be a problem in practice.

Examples

julia> R = exp(imx*π/4)  # Rotation through π/2 (note the extra 1/2) about the x axis
rotor(0.7071067811865476 + 0.7071067811865475𝐢 + 0.0𝐣 + 0.0𝐤)

julia> R * imx * conj(R)
0.0 + 1.0𝐢 + 0.0𝐣 + 0.0𝐤

julia> R * imy * conj(R)
0.0 + 0.0𝐢 + (2.220446049250313e-16)𝐣 + 1.0𝐤

julia> R * imz * conj(R)
0.0 + 0.0𝐢 - 1.0𝐣 + (2.220446049250313e-16)𝐤

Notes

The quaternionic exponential is very easy to calculate by analogy with the complex exponential. We can write a quaternion as $q = s + v\hat{v}$, where $s$ is the scalar part, $v$ is the norm of the pure-vector part, and $\hat{v}$ is a unit vector. Then, $\hat{v}^2 = -1$, so it acts exactly like the imaginary unit $i$ in complex numbers. Obviously, $s$, $v$, and $\hat{v}$ all commute with each other, so the math is simple and we can immediately calculate in analogy with Euler's formula for complex numbers:

\[\exp(q) = \exp(s) \left(\cos(v) + \hat{v}\sin(v)\right).\]

The only special case is when $v=0$, in which case there is no unique choice of $\hat{v}$, but then exp(q) = exp(s). Unfortunately, this means that we need to use a separate branch for this isolated case, which means that automatic differentiation with respect to the vector components will incorrectly produce a derivative of zero at that point if we do nothing but return the value. Therefore, we actually use a Taylor expansion to compute the result through fifth order in the magnitude of the vector part. This will produce correct derivatives up to fifth order.

Base.log — Method

log(q)

Logarithm of a quaternion.

Branch-cut behavior

As with the complex logarithm, the quaternion logarithm is multi-valued: you could add any integer multiple of $2πq̂$ (for some unit vector $q̂$) to the result of this function and get the same result after exponentiating. This function is the principal logarithm: the choice with the smallest-magnitude vector part.

Similarly, a branch cut is imposed along the negative real axis: if the vector components of q are precisely zero and the scalar component is negative, the returned quaternion will be log(-q[1]) + π𝐤. Unlike in the complex case, the choice of 𝐤 is arbitrary; the vector component of the returned quaternion could be π times any unit vector.

Automatic differentiation caveats

Values of q that are very close to (but not on) the negative real axis will produce accurate results, but automatic differentiation is likely to become numerically unstable. Precisely on the non-positive real axis, derivatives should not be defined at all, but will typically return incorrect (finite) values. These regions should be avoided in any case, because of the analytic discontinuities and geometric ambiguities.

Examples

julia> log(exp(1.2imy))
 + 0.0𝐢 + 1.2𝐣 + 0.0𝐤

julia> log(quaternion(exp(7)))
7.0 + 0.0𝐢 + 0.0𝐣 + 0.0𝐤

julia> log(quaternion(-exp(7)))
7.0 + 0.0𝐢 + 0.0𝐣 + 3.141592653589793𝐤

Notes

This function uses an accurate algorithm for finding the logarithm, although automatic differentiation may become inaccurate near the negative real axis because the analytic derivatives blow up there.

The log function is very analogous to the sqrt function [which is essentially exp(log(q)/2)], in that both functions have discontinuous behavior along the negative real axis, and derivatives that blow up as you approach that axis. Therefore, the same caveats about computing derivatives using automatic differentiation apply here as well. Specifically, those derivatives will be numerically unstable near the negative real axis, and will be incorrect on the negative real axis itself; rather than being undefined, the derivatives will typically be returned as zero because of our fixed choice of result there.

If we decompose the result of this function as $\log(q) = s + \vec{v}$, where $s$ is the scalar part and $\vec{v}$ is the pure-vector part, then clearly $s$ and $\vec{v}$ commute, so their exponentials also commute, and we have

\[q = \exp\left(\log(q)\right) = \exp\left(s\right) \exp\left(\vec{v}\right).\]

Note that the exponential of a pure-vector quaternion is a unit quaternion, so we have decomposed $q$ into a product of a positive real number and a unit quaternion. We already know how to take the logarithm of a positive real number. Therefore, in the following we assume that $q$ is a unit quaternion, so that $s=0$ and we only need to find $\vec{v}$.

We now write $\log(q)$ as $v\hat{v}$, where $v$ is just the scalar norm. Note that, because of the periodicity of the exp function, we can assume that $v \in [0, \pi]$ and, in particular, $\sin(v) \geq 0$. Now, expand the exponential as

\[\exp\left(\vec{v}\right) = \exp\left(v \hat{v}\right) = \cos(v) + \hat{v} \sin(v).\]

The input to this function is the right-hand side, but we do not yet know its decomposition into $v$ and $\hat{v}$. But we can find $\cos(v)$ as the scalar part of the input, and $\sin(v)$ as the absvec (since we know that $\sin(v) \geq 0$). Then, we can compute $v = \mathrm{atan}(\sin(v), \cos(v))$. And finally, we simply multiply the vector part of the input by $v / \sin(v)$ to obtain the logarithm. This factor is given accurately by invsinc(v) whenever $|v| \leq \pi/2$.

When that condition is not satisfied (which also implies $\cos(v)<0$, we can rewrite the problem as

\[\exp\left(v \hat{v}\right) = \cos(v) + \hat{v} \sin(v) = -\cos(v-\pi) - \hat{v} \sin(v-\pi) = -\cos(v') - \hat{v} \sin(v').\]

Here, we want to multiply the vector component by $-v / \sin(v') = -(v'+\pi) / \sin(v')$. Note that we can easily compute $v' = \mathrm{atan}(\sin(v), -\cos(v))$. This algorithm is surprisingly accurate, even when $v$ is extremely close to $\pi$, which implies that the vector part of the input is extremely small.

The only special case remaining to handle is when $\cos(v) < 0$ but $\sin(v)$ is identically zero. In this case, we could throw an error, but this is not usually helpful. Instead, we arbitrarily choose to return $\pi 𝐤$.

If q is a Rotor, we return a QuatVec; if q is a general Quaternion, we return a general Quaternion — though if q happens to have norm exactly 1, the result will have a scalar part of exactly 0. Note that, because there is no geometric reason to take the logarithm of a QuatVec, that case is not implemented; if you really need to compute it, you can convert the QuatVec to a Quaternion first.

Base.sqrt — Method

sqrt(q)

Square root of a quaternion.

Branch-cut behavior

As with the logarithm, the quaternionic square-root has a branch cut along the non-positive real axis: if the vector components of q are precisely zero and the scalar component is negative, the returned quaternion will be √(-q[1]) * 𝐤. Unlike in the complex case, the choice of 𝐤 is arbitrary; the vector component of the returned quaternion could be any unit vector.

Automatic differentiation caveats

Values of q that are very close to (but not on) the negative real axis will produce accurate results, but automatic differentiation is likely to become numerically unstable. Precisely on the non-positive real axis, derivatives should not be defined at all, but will typically return incorrect (finite) values. These regions should be avoided in any case, because of the analytic discontinuities and geometric ambiguities.

Examples

julia> q = quaternion(1.2, 3.4, 5.6, 7.8);

julia> sqrtq = √q;

julia> sqrtq^2 ≈ q
true

julia> √quaternion(4.0)
2.0 + 0.0𝐢 + 0.0𝐣 + 0.0𝐤

julia> √quaternion(-4.0)
0.0 + 0.0𝐢 + 0.0𝐣 + 2.0𝐤

Notes

The general formula whenever the denominator is nonzero is

\[\sqrt{q} = \frac{|q| + q} {\sqrt{2(|q| + q[1])}}\]

This can be proven by squaring the numerator, using q = q[1] + q⃗ and |q|^2 = q[1]^2 - q⃗^2 = q[1]^2 + |q⃗|^2. When the denominator is zero — or quite simply whenever q[1] < 0 so that the denominator may be subject to cancellation — we can use the fact that

\[|q| + q[1] = \frac{|q⃗|^2} {|q| − q[1]}\]

to evaluate the expression in a more stable form:

\[\sqrt{q} = \left(|q| + q\right)\sqrt{\frac{|q| − q[1]} {2|q⃗|^2}}.\]

Note that whenever the vector part is zero and the scalar part is negative, the solution is not unique (and the denominator above is zero), because it necessarily involves the square root of -1, of which there are infinitely many in the space of quaternions. In this case, we arbitrarily choose the vector part of the result to be proportional to 𝐤, as mentioned above. A reasonable alternative would be to throw an error; instead it is left to the user to check for that condition if it would be a problem.

Analytically, any derivative of this function will blow up as you approach the negative real axis, because the function is discontinuous there. Therefore, you should not expect to be able to accurately compute derivatives of this function at points near the negative real axis (including q=0) using automatic differentiation. Specifically on the negative real axis the derivative is not defined at all, but because of our fixed choice of result here, automatic differentiation will typically (and incorrectly) return the derivative as zero. Ultimately, the reason for this is geometrical, so it should be avoided by the user in any case. Therefore, for the sake of efficiency and accuracy when the problem is geometrically well conditioned, we do not attempt to special-case the derivative at these points.

Quaternionic.abs2vec — Method

abs2vec(q)

Sum of the squares of the three "vector" components of the quaternion.

Examples

julia> abs2vec(quaternion(1,2,3,6))
49

Quaternionic.absvec — Method

absvec(q)

Square root of the sum of the squares of the three "vector" components of the quaternion.

This function uses Julia's built-in hypot function to avoid overflow and underflow.

Examples

julia> absvec(quaternion(1,2,3,6))
7.0

Random quaternions

It is frequently convenient to construct random Quaternion objects, which can be done just as with other types by passing the desired output type to the randn function. The rand function is not overloaded, because there would be no geometric significance to such a Quaternion; randn results are independent of the orientation of the basis used to define the quaternions. Note that it is possible to get random rotors and vectors by passing the appropriate types to the randn function.

Base.randn — Method

randn([rng=default_rng()], T=Quaternion{Float64}, [dims...])

Generate a normally distributed random quaternion of type T with mean 0 and standard deviation of norm 1. Optionally generate an array of such quaternions. This module currently provides an implementation for the types QuaternionF16, QuaternionF32, and QuaternionF64 (the default). The values are drawn from the spherically symmetric quaternionic normal distribution of variance 1 (corresponding to each component having independent normal distribution with mean zero and variance 1/4).

Note that this function works with Quaternion{BigFloat}, even though Base.randn does not work with BigFloat on Julia <1.9; for earlier versions, we just use the Box-Muller transform to obtain the desired result.

If the quaternion type passed in is Rotor, the result will be normalized correctly. Because the distribution is spherically symmetric, the result is a truly random rotation.

If the quaternion type is QuatVec, the result will have a 0 scalar component, and the vector will have mean 0 standard deviation of norm 1.

Examples

julia> randn(QuaternionF64)
0.4336736009756228 - 0.45087190792840853𝐢 - 0.24723937675211696𝐣 - 0.4514571469326208𝐤
julia> randn(QuaternionF16, 2, 2)
2×2 Matrix{QuaternionF16}:
   0.4321 + 1.105𝐢 + 0.2664𝐣 - 0.1359𝐤   0.064 + 0.9263𝐢 - 0.4138𝐣 + 0.05505𝐤
 0.2512 - 0.2585𝐢 - 0.2803𝐣 - 0.00964𝐤  -0.1256 + 0.1848𝐢 + 0.03607𝐣 - 0.752𝐤

Conversions

It can sometimes be useful to convert between quaternions and other representations. Most of these functions are named to_<representation> and have a corresponding from_<representation> function. Furthermore, most convert to/from representations of rotations. While rotations are not the only useful application of quaternions, they are probably the most common. The only conversions that are not specifically related to rotations are to_float_array and from_float_array.

Quaternionic.from_euler_angles — Method

from_euler_angles(α, β, γ)

Come over from the dark side.

Assumes the Euler angles correspond to the quaternion R via

R = exp(α𝐤/2) * exp(β𝐣/2) * exp(γ𝐤/2)

where 𝐣 and 𝐤 rotate about the fixed $y$ and $z$ axes, respectively, so this reprents an initial rotation (in the positive sense) through an angle $γ$ about the axis $z$, followed by a rotation through $β$ about the axis $y$, and a final rotation through $α$ about the axis $z$. This is equivalent to performing an initial rotation through $α$ about the axis $z$, followed by a rotation through $β$ about the rotated axis $y'$, followed by a rotation through $γ$ about the twice-rotated axis $z''$. The angles are naturally assumed to be in radians.

NOTE: Before opening an issue reporting something "wrong" with this function, be sure to read all of this page, especially the very last section about opening issues or pull requests.

See Also

to_euler_angles: Convert quaternion to Euler angles

to_euler_phases: Convert quaternion to Euler phases

from_euler_phases: Create quaternion from Euler phases

Quaternionic.from_euler_phases — Method

from_euler_phases(zₐ, zᵦ, zᵧ)
from_euler_phases(z)

Return the quaternion corresponding to these Euler phases.

Interpreting the input quaternion as a rotation (though its normalization scales out), we can define the complex Euler phases from the Euler angles (α, β, γ) as

zₐ ≔ exp(i*α)
zᵦ ≔ exp(i*β)
zᵧ ≔ exp(i*γ)

These are more useful geometric quantites than the angles themselves — being involved in computing spherical harmonics and Wigner's 𝔇 matrices — and can be computed from the components of the corresponding quaternion algebraically (without the use of transcendental functions).

Parameters

z::Vector{Complex{T}}: complex vector of length 3, representing the complex phases (zₐ, zᵦ, zᵧ) in that order.

Returns

R::Quaternion{T}

See Also

to_euler_phases: Convert quaternion to Euler phases

to_euler_angles: Convert quaternion to Euler angles

from_euler_angles: Create quaternion from Euler angles

Quaternionic.from_float_array — Method

from_float_array(A)

Reinterpret a float array as an array of quaternions

The input array must have an initial dimension whose size is 4, because successive indices in that dimension will be considered successive components of the output quaternion.

Note that this returns a view of the original data [via reinterpret(reshape,...)] only if the base type of the input array isbitstype; otherwise, a new array of Quaternions must be created, and the memory copied.

See also to_float_array.

Quaternionic.from_rotation_matrix — Method

from_rotation_matrix(ℛ)

Convert 3x3 rotation matrix to quaternion.

Assuming the 3x3 matrix ℛ rotates a vector v according to

v' = ℛ * v,

we can also express this rotation in terms of a quaternion R such that

v' = R * v * R⁻¹.

This function returns that quaternion, using Bar-Itzhack's algorithm (version 3) to allow for non-orthogonal matrices. J. Guidance, Vol. 23, No. 6, p. 1085

Note

If you want to use this function for matrices with elements of types other than Float64 or Float32, you will need to (install and) import GenericLinearAlgebra first. The reason is that this function computes the eigen-decomposition of ℛ, which is only available for more generic float types via that package. Note that you will want at least version 0.3.11 of GenericLinearAlgebra because previous versions had a bug.

Quaternionic.from_spherical_coordinates — Method

from_spherical_coordinates(θ, ϕ)

Return a rotor corresponding to these spherical coordinates.

Considering (θ, ϕ) as a point $n̂$ on the sphere, we can also construct a quaternion that rotates the $z$ axis onto that point. Here, we use the standard commonly used in physics: θ represents the "polar angle" between the $z$ axis and the direction $n̂$, while ϕ represents the "azimuthal angle" between the $x$ axis and the projection of $n̂$ into the $x$-$y$ plane. Both angles must be given in radians.

Quaternionic.to_euler_angles — Method

to_euler_angles(R)

Open Pandora's Box.

If somebody is trying to make you use Euler angles, tell them no, and walk away, and go and tell your mum.

You don't want to use Euler angles. They are awful. Stay away. It's one thing to convert from Euler angles to quaternions; at least you're moving in the right direction. But to go the other way?! It's just not right.

Assumes the Euler angles correspond to the quaternion R via

R = exp(α𝐤/2) * exp(β𝐣/2) * exp(γ𝐤/2)

where 𝐣 and 𝐤 rotate about the fixed $y$ and $z$ axes, respectively, so this represents an initial rotation (in the positive sense) through an angle $γ$ about the axis $z$, followed by a rotation through $β$ about the axis $y$, and a final rotation through $α$ about the axis $z$. This is equivalent to performing an initial rotation through $α$ about the axis $z$, followed by a rotation through $β$ about the rotated axis $y'$, followed by a rotation through $γ$ about the twice-rotated axis $z''$. The angles are naturally assumed to be in radians.

The outputs from this function are in these ranges:

α ∈ (-2π, 2π]

β ∈ [0, π]

γ ∈ (-2π, 2π]

This is redundant, and inconsistent with more standard conventions for Euler angles on $\mathrm{Spin}(3)$:

α ∈ [0, 2π)

β ∈ [0, π]

γ ∈ [0, 4π)

But since these angles will usually be fed into periodic functions, it usually won't matter. If you really need angles in those ranges, you can always post-process the output of this function.

NOTE: Before opening an issue reporting something "wrong" with this function, be sure to read all of this page, especially the very last section about opening issues or pull requests.

Returns

αβγ::StaticVector{T}

Raises

AllHell if you try to actually use Euler angles, when you could have been using quaternions like a sensible person.

See Also

from_euler_angles: Create quaternion from Euler angles

to_euler_phases: Convert quaternion to Euler phases

from_euler_phases: Create quaternion from Euler phases

Quaternionic.to_euler_phases — Method

to_euler_phases(q)

Convert input quaternion to complex phases of Euler angles

Interpreting the input quaternion as a rotation (though its normalization scales out), we can define the complex Euler phases from the Euler angles (α, β, γ) as

zₐ ≔ exp(i*α)
zᵦ ≔ exp(i*β)
zᵧ ≔ exp(i*γ)

These are more useful geometric quantities than the angles themselves — being involved in computing spherical harmonics and Wigner's 𝔇 matrices — and can be computed from the components of the corresponding quaternion algebraically (without the use of transcendental functions).

Note that to_euler_phases!(z, q) is supported for backwards compatibility, but because this function returns an SVector, there is probably no advantage to the in-place approach.

Incorrect derivatives near singularities

Note that the derivatives of these phases with respect to the quaternion components are ill-defined at the Euler singularities (specifically, when β=0 or β=π). The problematic cases will just return 1 for the phase factor, which will autodiff to zero. Fortunately, this should not pollute downstream results if they are grounded in geometric quantities. Specifically, Wigner's 𝔇 matrices can be computed from these phases without any issues, because the $d$ factors should be zero anyway, and the product rule will save us.

Returns

z::SVector{Complex{T}}: complex phases (zₐ, zᵦ, zᵧ) in that order.

See Also

from_euler_phases: Create quaternion from Euler phases

to_euler_angles: Convert quaternion to Euler angles

from_euler_angles: Create quaternion from Euler angles

Quaternionic.to_float_array — Method

to_float_array(A)

View a quaternion array as an array of numbers

The output array will have an extra initial dimension whose size is 4, because successive indices in that dimension correspond to successive components of the quaternion.

Note that this returns a view of the original data only if the base type of the input array isbitstype; otherwise, a new array of that type must be created, and the memory copied.

See also from_float_array.

Quaternionic.to_rotation_matrix — Method

to_rotation_matrix(q)

Convert quaternion to 3x3 rotation matrix.

Assuming the quaternion R rotates a vector v according to

v' = R * v * R⁻¹,

we can also express this rotation in terms of a 3x3 matrix ℛ such that

v' = ℛ * v.

This function returns that matrix.

Quaternionic.to_spherical_coordinates — Method

to_spherical_coordinates(q)

Return the spherical coordinates corresponding to this quaternion.

We can treat the quaternion as a transformation taking the $z$ axis to some direction $n̂$. This direction can be described in terms of spherical coordinates (θ, ϕ). Here, we use the standard commonly used in physics: θ represents the "polar angle" between the $z$ axis and the direction $n̂$, while ϕ represents the "azimuthal angle" between the $x$ axis and the projection of $n̂$ into the $x$-$y$ plane. Both angles are given in radians.

Distances

There are several ways of measuring the "distance" between two quaternions: $d(q_1, q_2)$. Fundamentally, any comparison between two quaternions $q_1$ and $q_2$ must make use of a binary operation, for which there are two obvious choices: addition or multiplication. For either choice, we operate on $q_1$ and the appropriate inverse (either additive or multiplicative) of $q_2$. That is, $d$ should be a function of either $q_1 - q_2$ or $q_1/q_2$.[1]

Now, we also have a number of criteria we would like any distance function to satisfy. For any quaternions $q_1$ and $q_2$ and any unit quaternion $q_3$, we require

real-valued: $d(q_1, q_2) \in \mathbb{R}$

symmetry: $d(q_1, q_2) = d(q_2, q_1)$

invariance: $d(q_3\, q_1, q_3\, q_2) = d(q_1, q_2) = d(q_1\, q_3, q_2\, q_3)$

identity: $d(q_1, q_1) = 0$

positive-definiteness: $d(q_1, q_2) > 0$ whenever $q_1 ≠ q_2$

(Of course, it should be noted that these criteria all hold in the exact case; when using floating-point numbers, they will likely be violated near edge cases.)

It is not hard to see that these criteria can be satisfied by any of

abs(q₁ - q₂)

abs2(q₁ - q₂)

abs(log(q₁ / q₂))

abs2(log(q₁ / q₂)

If $q_1$ and $q_2$ are interpreted as rotations, we frequently don't care about their signs, and just want the smallest distance between them, for any choice of sign. Furthermore, in the multiplicative case, the log functions will involve calculation of the log of the magnitudes of the quaternions, which should be 1. In this case, we relax the "positive-definiteness" criterion to allow $d(q_1, q_2)$ to equal zero when $q_1$ and $q_2$ are related by a nonzero scalar multiple.

For Rotor types, the latter two multiplicative options are most relevant, while for other types the additive options are more relevant. These are the default behaviors of the distance and distance2 functions.

Quaternionic.distance — Method

distance(q₁, q₂)
distance2(q₁, q₂)

Measure the "distance" between two quaternions, or the squared distance with distance2.

By default, this function just returns the natural measure in the additive group of quaternions:

abs2(q₁ - q₂)

If both arguments are Rotors, the function returns the natural measure in the rotation group, which is roughly

abs2(log(q₁ / q₂))

Note that for Rotors, this method is (efficiently) independent of the scaling of q₁ and q₂, including up to factors of -1, as is appropriate for the rotation group.

Examples

julia> distance(imx, imy)
1.4142135623730951
julia> distance(rotor(imx), rotor(imy))
1.5707963267948966
julia> distance(imz, -imz)
2.0
julia> distance(rotor(imz), rotor(-imz))
0.0

Alignment

There are many ways to optimize alignment with rotations. In particular, we can seek the optimal rotation that takes one set of points onto a corresponding set of points, or the optimal quaternion that takes one set of quaternions onto a corresponding set of quaternions. In both cases, the "optimal" value depends on the metric being used. The simplest and most robust results are obtained when the metric is the standard Euclidean metric (in the case of points), or the magnitude of the difference (in the case of quaternions). Here, we assume that QuatVecs represent points, and any other type of quaternion should be treated as rotors.

Quaternionic.align — Method

align(a⃗, b⃗, [w])

Solve Wahba's problem, finding a rotation that aligns the set of points a⃗ to a corresponding set of points b⃗ by minimizing the distance between the first set and the rotated second set.

Here, a⃗ and b⃗ must be equally sized arrays of QuatVecs. If present, w must be an equally sized array of real numbers; if not, it is taken to be 1. We define the loss function

\[L(ℛ) ≔ Σᵢ wᵢ ‖a⃗ᵢ - ℛ b⃗ᵢ‖²\]

where $ℛ$ is a rotation operator, and return the quaternion corresponding to the optimal $ℛ$ that minimizes this function.

Note that it is possible that the points do not uniquely determine a rotation — as when one or both sets of points is rotationally symmetric. In that case, the loss function $L(ℛ)$ will still be minimized and the points will still be optimally aligned by the output quaternion, but that quaternion will not be unique.

Notes

In their book Fundamentals of Spacecraft Attitude Determination and Control (2014), Markley and Crassidis say that "Davenport’s method remains the best method for solving Wahba’s problem". This method provides the optimal quaternion as the dominant eigenvector (the one with the largest eigenvalue) of a certain matrix. We start by defining the supplementary matrix

\[S ≔ Σᵢ wᵢ a⃗ᵢ b⃗ᵢᵀ\]

and vector

\[s⃗ ≔ \begin{bmatrix} S₂₃-S₃₂ \\ S₃₁-S₁₃ \\ S₁₂-S₂₁ \end{bmatrix}.\]

Then the key matrix is

\[M ≔ \begin{bmatrix} S + Sᵀ - (\mathrm{tr}S)\, I₃ & s⃗ \\ s⃗ᵀ & \mathrm{tr}S \end{bmatrix}\]

It is possible for this matrix to have degenerate eigenvalues, corresponding to cases where the points do not uniquely determine the rotation, as described above.

Quaternionic.align — Method

align(A, B, [w])

Find a Rotor that aligns the set of rotors A to a corresponding set B by minimizing the distance between the first set and the rotated second set.

Here, A and B must be equally sized arrays of AbstractQuaternions. If present, w must be an equally sized array of real numbers; if not, it is taken to be 1. We define the loss function

\[L(R) ≔ Σᵢ wᵢ |Aᵢ - R Bᵢ|²\]

where $R$ is a Rotor, and return the quaternion corresponding to the optimal $R$ that minimizes this function.

Note that it is possible that the input data do not uniquely determine a rotor, which will happen when sum below is zero. When this happens, the result will contain NaNs, but no error will be raised. When the sum is very close to — but not exactly — zero, the accuracy of the result will be limited. However, the loss function will not depend strongly on the result in that case.

Be aware that this function is sensitive to the signs of the input quaternions. See the unflip function for one way to avoid problems related to signs.

Notes

We can ensure that the loss function is minimized by multiplying $R$ by an exponential, differentiating with respect to the argument of the exponential, and setting that argument to 0. This derivative should be 0 at the minimum. We have

\[∂ⱼ Σᵢ wᵢ |Aᵢ - \exp[vⱼ] R Bᵢ|² → 2 ⟨ eⱼ R Σᵢ wᵢ Bᵢ Āᵢ ⟩₀\]

where → denotes taking $vⱼ→0$, the symbol $⟨⟩₀$ denotes taking the scalar part, and $eⱼ$ is the unit quaternionic vector in the $j$ direction. The only way for this quantity to be zero for each choice of $j$ is if

\[R Σᵢ wᵢ Bᵢ Āᵢ\]

is itself a pure scalar. This, in turn, can only happen if either (1) the sum is 0 or (2) if $R$ is proportional to the conjugate of the sum:

\[R ∝ Σᵢ wᵢ Aᵢ B̄ᵢ\]

Now, since we want $R$ to be a rotor, we simply define it to be the normalized sum.

1For $q_1/q_2$, we are dealing with the multiplicative group of quaternions, which does not include 0, so we will assume that no quaternion involved in such a function can be 0.